Refactor check_user_privileges to static methods
Refactor check_user_privileges.lib.php functions to static methods and move logic to check_user_privileges.inc.php.
