+++ libraries/classes/Navigation/Nodes/NodeDatabaseContainer.php
@@ -10,7 +10,7 @@ namespace PhpMyAdmin\Navigation\Nodes;
 use PhpMyAdmin\Navigation\NodeFactory;
 use PhpMyAdmin\Util;
 
-require_once './libraries/check_user_privileges.lib.php';
+require_once './libraries/check_user_privileges.inc.php';
 
 /**
  * Represents a container for database nodes in the navigation tree
