+++ server_privileges.php
@@ -22,7 +22,7 @@ require_once 'libraries/common.inc.php';
 /**
  * functions implementation for this script
  */
-require_once 'libraries/check_user_privileges.lib.php';
+require_once 'libraries/check_user_privileges.inc.php';
 
 $cfgRelation = Relation::getRelationsParam();
