+++ sql.php
@@ -17,7 +17,7 @@ use PhpMyAdmin\Util;
  * Gets some core libraries
  */
 require_once 'libraries/common.inc.php';
-require_once 'libraries/check_user_privileges.lib.php';
+require_once 'libraries/check_user_privileges.inc.php';
 
 PageSettings::showGroup('Browse');
